+++ apps/interactor/interactor/__init__.py
@@ -1,4 +1,4 @@
-VERSION = "0.9.1"
+VERSION = "0.10.0"
 
 
 def run_pytest():